Document public API for kTaglib
