+++ lib/src/main/java/com/simplecityapps/ktaglib/AudioFile.kt
@@ -1,5 +1,13 @@
 package com.simplecityapps.ktaglib
 
+/**
+ * A ready to use data model with some basic fields for representing
+ * various metadata fields of an audio file. If this data class is
+ * insufficient to meet your needs, directly use [KTagLib.getMetadata]
+ * to obtain a HashMap of all properties and write a mapping function
+ * to convert it to your model. For an example, see the source of
+ * [KTagLib.getAudioFile].
+ */
 data class AudioFile(
     val path: String,
     var size: Long,
