+++ sample/src/main/java/com/simplecityapps/ktaglib/sample/MainActivity.kt
@@ -12,13 +12,11 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.recyclerview.widget.RecyclerView
 import com.simplecityapps.ktaglib.AudioFile
 import com.simplecityapps.ktaglib.KTagLib
-import com.simplecityapps.ktaglib.TaglibUtils
 import kotlinx.coroutines.*
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.flow.flow
 import kotlinx.coroutines.flow.flowOn
-import java.util.HashMap
 
 class MainActivity : AppCompatActivity() {
 
@@ -29,9 +27,6 @@ class MainActivity : AppCompatActivity() {
     private val scope = CoroutineScope(Dispatchers.Main + exceptionHandler)
 
     private lateinit var documentAdapter: DocumentAdapter
-
-    private val tagLib = KTagLib()
-
 
     // Lifecycle
 
@@ -131,7 +126,7 @@ class MainActivity : AppCompatActivity() {
             documents.forEach { document ->
                 contentResolver.openFileDescriptor(document.uri, "r")?.use { pfd ->
                     try {
-                        emit(Pair(TaglibUtils.getAudioFile(pfd.fd, document.uri.toString(), document.displayName.substringBeforeLast(".")), document))
+                        emit(Pair(KTagLib.getAudioFile(pfd.fd, document.uri.toString(), document.displayName.substringBeforeLast(".")), document))
                     } catch (e: IllegalStateException) {
                         Log.e("MainActivity", "Failed to get audio file: ", e)
                     }
